renaming and fixing for files
