+++ app/src/main/java/com/catnip/rizkyilmann_challange4/ui/cart/CartFragment.kt
@@ -1,4 +1,4 @@
-package com.catnip.rizkyilmann_challange4.ui.dashboard
+package com.catnip.rizkyilmann_challange4.ui.cart
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.catnip.rizkyilmann_challange4.databinding.FragmentDashboardBinding
+import com.catnip.rizkyilmann_challange4.databinding.FragmentCartBinding
 
-class DashboardFragment : Fragment() {
+class CartFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentCartBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -23,9 +23,9 @@ class DashboardFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+            ViewModelProvider(this).get(CartViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentCartBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
         val textView: TextView = binding.textDashboard
