+++ app/src/main/java/com/catnip/rizkyilmann_challange4/ui/cart/CartViewModel.kt
@@ -1,13 +1,13 @@
-package com.catnip.rizkyilmann_challange4.ui.notifications
+package com.catnip.rizkyilmann_challange4.ui.cart
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class CartViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
