+++ app/src/main/java/com/catnip/rizkyilmann_challange4/ui/profile/ProfileViewModel.kt
@@ -1,13 +1,13 @@
-package com.catnip.rizkyilmann_challange4.ui.dashboard
+package com.catnip.rizkyilmann_challange4.ui.profile
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class ProfileViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
