+++ app/src/main/res/drawable/ic_user.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="33dp"
+    android:height="33dp"
+    android:viewportWidth="33"
+    android:viewportHeight="33">
+  <path
+      android:pathData="M16.5,30.25C14.421,30.256 12.369,29.785 10.5,28.875C9.816,28.542 9.16,28.154 8.54,27.713L8.352,27.576C6.647,26.317 5.252,24.684 4.276,22.803C3.267,20.856 2.743,18.693 2.75,16.5C2.75,8.906 8.906,2.75 16.5,2.75C24.094,2.75 30.25,8.906 30.25,16.5C30.257,18.692 29.734,20.854 28.725,22.8C27.75,24.681 26.358,26.313 24.655,27.572C24.013,28.042 23.331,28.456 22.618,28.809L22.507,28.864C20.637,29.779 18.582,30.253 16.5,30.25ZM16.5,23.375C14.44,23.371 12.551,24.522 11.609,26.355C14.691,27.881 18.309,27.881 21.391,26.355V26.348C20.448,24.517 18.559,23.369 16.5,23.375ZM16.5,20.625C19.479,20.629 22.225,22.234 23.69,24.827L23.711,24.809L23.73,24.793L23.706,24.813L23.693,24.824C27.17,21.82 28.413,16.971 26.811,12.664C25.208,8.357 21.097,5.5 16.501,5.5C11.906,5.5 7.795,8.357 6.192,12.664C4.589,16.971 5.833,21.82 9.31,24.824C10.776,22.232 13.522,20.629 16.5,20.625ZM16.5,19.25C13.463,19.25 11,16.788 11,13.75C11,10.712 13.463,8.25 16.5,8.25C19.538,8.25 22,10.712 22,13.75C22,15.209 21.421,16.608 20.389,17.639C19.358,18.67 17.959,19.25 16.5,19.25ZM16.5,11C14.981,11 13.75,12.231 13.75,13.75C13.75,15.269 14.981,16.5 16.5,16.5C18.019,16.5 19.25,15.269 19.25,13.75C19.25,12.231 18.019,11 16.5,11Z"
+      android:fillColor="#2E3A59"/>
+</vector>
