+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notifications.NotificationsFragment"
+    tools:context=".ui.profile.ProfileFragment"
     android:background="@color/yellow1">
 
     <androidx.cardview.widget.CardView
@@ -80,7 +80,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
                 android:layout_marginTop="25dp"
-                android:text="NAMA"
+                android:text="@string/tv_nama_profile"
                 android:textColor="@color/black"
                 app:layout_constraintStart_toEndOf="@id/iv_profile_name"
                 app:layout_constraintTop_toTopOf="@id/iv_profile_name" />
@@ -133,7 +133,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
                 android:layout_marginTop="25dp"
-                android:text="EMAIL"
+                android:text="@string/tv_email_profile"
                 android:textColor="@color/black"
                 app:layout_constraintStart_toEndOf="@id/iv_profile_email"
                 app:layout_constraintTop_toTopOf="@id/iv_profile_email" />
@@ -186,7 +186,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
                 android:layout_marginTop="25dp"
-                android:text="No. Telephone"
+                android:text="@string/text_nomortelepon"
                 android:textColor="@color/black"
                 app:layout_constraintStart_toEndOf="@id/iv_profile_number"
                 app:layout_constraintTop_toTopOf="@id/iv_profile_number" />
