+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.catnip.rizkyilmann_challange4.ui.dashboard.DashboardFragment"
+        android:name="com.catnip.rizkyilmann_challange4.ui.cart.CartFragment"
         android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        tools:layout="@layout/fragment_cart" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.catnip.rizkyilmann_challange4.ui.notifications.NotificationsFragment"
+        android:name="com.catnip.rizkyilmann_challange4.ui.profile.ProfileFragment"
         android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        tools:layout="@layout/fragment_profile" />
 </navigation>
